+++ packages/node/src/indexer/blockDispatcher/worker-block-dispatcher.service.ts
@@ -6,7 +6,6 @@ import { Inject, Injectable, OnApplicationShutdown } from '@nestjs/common';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import {
   NodeConfig,
-  SmartBatchService,
   StoreService,
   StoreCacheService,
   IProjectService,
@@ -48,7 +47,6 @@ export class WorkerBlockDispatcherService
     projectService: IProjectService<SubstrateDatasource>,
     @Inject('IProjectUpgradeService')
     projectUpgadeService: IProjectUpgradeService,
-    smartBatchService: SmartBatchService,
     cacheService: InMemoryCacheService,
     storeService: StoreService,
     storeCacheService: StoreCacheService,
@@ -63,12 +61,10 @@ export class WorkerBlockDispatcherService
       eventEmitter,
       projectService,
       projectUpgadeService,
-      smartBatchService,
       storeService,
       storeCacheService,
       poiSyncService,
       project,
-      dynamicDsService,
       () =>
         createIndexerWorkerCore<
           IIndexerWorker,